atomics (erts v12.0.3)

This module provides a set of functions to do atomic operations towards mutable atomic variables. The implementation utilizes only atomic hardware instructions without any software level locking, which makes it very efficient for concurrent access. The atomics are organized into arrays with the following semantics:

Atomics are 64 bit integers.

Atomics can be represented as either signed or unsigned.

Atomics wrap around at overflow and underflow operations.

All operations guarantee atomicity. No intermediate results can be seen. The result of one mutation can only be the input to one following mutation.

All atomic operations are mutually ordered. If atomic B is updated after atomic A, then that is how it will appear to any concurrent readers. No one can read the new value of B and then read the old value of A.

Indexes into atomic arrays are one-based. An atomic array of arity N contains N atomics with index from 1 to N.

Types

-type atomics_ref() :: term().
(opaque)

Specs

atomics_ref()

Identifies an atomic array returned from new/2.

Functions

add/3
(since OTP 21.2)

Specs

add(Ref, Ix, Incr) -> ok when Ref :: atomics_ref(), Ix :: integer(), Incr :: integer().

Add Incr to atomic.

add_get/3
(since OTP 21.2)

Specs

add_get(Ref, Ix, Incr) -> integer() when Ref :: atomics_ref(), Ix :: integer(), Incr :: integer().

Atomic addition and return of the result.

compare_exchange/4
(since OTP 21.2)

Specs

compare_exchange(Ref, Ix, Expected, Desired) -> ok | integer()
                    when
                        Ref :: atomics_ref(),
                        Ix :: integer(),
                        Expected :: integer(),
                        Desired :: integer().

Atomically compares the atomic with Expected, and if those are equal, set atomic to Desired. Returns ok if Desired was written. Returns the actual atomic value if not equal to Expected.

exchange/3
(since OTP 21.2)

Specs

exchange(Ref, Ix, Desired) -> integer()
            when Ref :: atomics_ref(), Ix :: integer(), Desired :: integer().

Atomically replaces the value of the atomic with Desired and returns the value it held previously.

get/2
(since OTP 21.2)

Specs

get(Ref, Ix) -> integer() when Ref :: atomics_ref(), Ix :: integer().

Read atomic value.

info/1
(since OTP 21.2)

Specs

info(Ref) -> Info
        when
            Ref :: atomics_ref(),
            Info :: #{size := Size, max := Max, min := Min, memory := Memory},
            Size :: non_neg_integer(),
            Max :: integer(),
            Min :: integer(),
            Memory :: non_neg_integer().

Return information about an atomic array in a map. The map has the following keys:

size

The number of atomics in the array.

max

The highest possible value an atomic in this array can hold.

min

The lowest possible value an atomic in this array can hold.

memory

Approximate memory consumption for the array in bytes.

new/2
(since OTP 21.2)

Specs

new(Arity, Opts) -> atomics_ref()
       when Arity :: pos_integer(), Opts :: [Opt], Opt :: {signed, boolean()}.

Create a new array of Arity number of atomics. All atomics in the array are initially set to zero.

Argument Opts is a list of the following possible options:

{signed, boolean()}

Indicate if the elements of the array will be treated as signed or unsigned integers. Default is true (signed).

The integer interval for signed atomics are from -(1 bsl 63) to (1 bsl 63)-1 and for unsigned atomics from 0 to (1 bsl 64)-1.

Atomics are not tied to the current process and are automatically garbage collected when they are no longer referenced.

put/3
(since OTP 21.2)

Specs

put(Ref, Ix, Value) -> ok when Ref :: atomics_ref(), Ix :: integer(), Value :: integer().

Set atomic to Value.

sub/3
(since OTP 21.2)

Specs

sub(Ref, Ix, Decr) -> ok when Ref :: atomics_ref(), Ix :: integer(), Decr :: integer().

Subtract Decr from atomic.

sub_get/3
(since OTP 21.2)

Specs

sub_get(Ref, Ix, Decr) -> integer() when Ref :: atomics_ref(), Ix :: integer(), Decr :: integer().

Atomic subtraction and return of the result.